week03/onclass/n_ary_tree_level_order_traversal_test.go
package week03onclass

import (
	"finger2011/algggorithm/internel"
	"reflect"
	"testing"
)

func Test_levelOrder(t *testing.T) {
	type args struct {
		root *internel.NodeTree
	}
	tests := []struct {
		name string
		args args
		want [][]int
	}{
		// TODO: Add test cases.
		{
			name: "test01",
			args: args{root: internel.CreateNodeTree([]int{1, internel.MinInt, 3, 2, 4, internel.MinInt, 5, 6})},
			want: [][]int{[]int{1}, []int{3, 2, 4}, []int{5, 6}},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := levelOrder(tt.args.root); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("levelOrder() = %v, want %v", got, tt.want)
			}
		})
	}
}
